refactor besu a bit, fix flaw with missing roots on invalid tx
